pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-sql-review-env"
version = "1.0.0"
description = "SQL Query Review environment for OpenEnv — fix, optimize, and secure SQL queries"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.57.4",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["sql_review_env", "sql_review_env.server"]
package-dir = { "sql_review_env" = ".", "sql_review_env.server" = "server" }
